requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.25.1
pydantic>=2.4.2
pydantic-settings>=2.1.0
pydantic[email]>=2.4.2
python-dotenv>=1.0.0
prisma>=0.11.0
python-multipart>=0.0.6
email-validator>=2.1.0.post1
apscheduler>=3.10.4
bcrypt>=4.0.0
